requirement.txt
autopep8==1.6.0
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
click==7.1.2
cryptography==3.3.2
decorator==4.4.2
distlib==0.3.4
emoji==0.5.4
filelock==3.5.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
instagram-private-api==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
platformdirs==2.5.0
pycodestyle==2.8.0
pycparser==2.20
pyTelegramBotAPI==3.7.1
python-decouple==3.3
python-dotenv==0.14.0
python-slugify==3.0.2
python-telegram-bot==12.8
requests==2.24.0
six==1.15.0
text-unidecode==1.2
toml==0.10.2
tornado==6.0.4
urllib3==1.25.11
virtualenv==20.13.1
Werkzeug==1.0.1
